Fixed rename errors
